Used tailwind in Home Page and removed the css file and moved it out of the directory.
