--- src/pages/Home/index.css
@@ -1,37 +1,0 @@
-.home-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: 100vh;
-  background-color: var(--clr-primary-light);
-}
-
-.content {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-  background-color: white;
-  padding: 3em;
-  border-radius: 20px;
-  box-shadow: rgba(0, 0, 0, 0.2) 0px 15px 30px;
-  max-width: 500px;
-}
-
-.home-title {
-  font-size: var(--primary-heading-font-size);
-  color: var(--clr-primary-dark);
-  margin: 20px 0 10px;
-}
-
-.home-subtitle {
-  font-size: var(--secondary-font-size);
-  color: var(--clr-secondary-dark);
-  margin-bottom: 20px;
-}
-
-.home-buttons {
-  display: flex;
-  gap: 1.5em;
-  margin-top: 20px;
-}
